Detail on changes:
Renamed name into tag. This is because I am planning to use labelComponent to show the number of gate each component has. Everything should be compatible but please check for bugs
calculation logic for each gates just added "state.NotConnected" condition. Consider it as an error check and should not cause bugs (please check for bugs)